Fix issue with `has_attribute?` validation
When there is a model that is not yet backed by a table, an error will
be raised when calling `has_attribute?`.  This isn't really an issue if
there is a pending migration.  But if there is not a pending migration,
something is fishy as the Enumbler should be backed by a database table.